// nodejs/src/types.ts
export interface StreamChunkMessage {
  type: "stream_chunk";
  from: string;
  chunkType: string;
  data: unknown;
}

export interface StreamEndMessage {
  type: "stream_end";
  from: string;
  messageId?: string;
}

export interface StreamCancelMessage {
  type: "stream_cancel";
  from: string;
}